Renamed the visitor attribute, separated out the visitor source generator part
